Bug fixes.
Adds more advance calculation of the viewport width depending on the amount of the slides and of the requested slides to show on a screen. Adds cloning of the first and last slides for smooth transition, when there is a small number of slides, e.g 3 slides and 3 slides to show on the screen.